building the basis for a turned based text based game but I need to implement stacks. the cooldowns will help since I can add the skills to a list and i can use peek(); to see if the previous skill is still on cooldown but I need to use push and .isEmpty(); maybe i can implement in moving around the map
9/14/2024:
Setting up monster and player skills
implemented interfaces for modularity,
each skill is responsible for its own methods leading to more robust skills
added to other base classes, but nothing notable
 currently building combat logic to stack skills for cooldown need to implement peek():
will implement REDO ATTACK DEFEND SKILL EQUIP will need to place skills into an array that I can select from
will work on weapons class next which will add buffs to player attributes based off item
TODO:
need to build simple map and use stack logic to move from one area to the next
need to create random encounters
need to finish weapons class
need to create equipment
need to create action list??

stopping for the night finished basic combat but it is still broken

worked on movement and setting up for special events
need to build special events
created run function but need to make better
can look at map to see where player it

combat is less broken need to fix issue with monster attacking multiple times and if i uise the other code i exit right after I attack

fixed issue of re-instantiating player a second time when maze is created

currently runs but I am facing issues where turns will still execute even after death.
need to still refactor turn based fight logic.
