+++ src/main/java/dtcc/itn262/app/Main.java
@@ -1,6 +1,6 @@
-package dtcc.itn262;
+package dtcc.itn262.app;
 
-import dtcc.itn262.gameutilities.Game;
+import dtcc.itn262.utilities.gamecore.Game;
 
 public class Main {
     public static void main(String[] args) {
