+++ src/main/java/dtcc/itn262/combat/MonsterActions.java
@@ -1,7 +1,7 @@
 package dtcc.itn262.combat;
 
 import dtcc.itn262.character.Player;
-import dtcc.itn262.gameutilities.Constants;
+import dtcc.itn262.utilities.gamecore.Constants;
 import dtcc.itn262.monster.Monster;
 import dtcc.itn262.monster.MonsterAttributes;
 import dtcc.itn262.skills.MonsterSkill;
