+++ src/main/java/dtcc/itn262/utilities/gamecore/Command.java
@@ -1,4 +1,4 @@
-package dtcc.itn262.gameutilities;
+package dtcc.itn262.utilities.gamecore;
 
 public enum Command {
     MAP, PROGRESS, HELP, EXIT, NORTH, SOUTH, EAST, WEST, HISTORY
